+++ pkg/resolvconf/resolv_windows.go
@@ -1,0 +1,13 @@
+package resolvconf
+
+func AddNameserver(nameserver string) error {
+	return nil
+}
+
+func RestoreConfig() error {
+	return nil
+}
+
+func GetLocalDomains() string {
+	return ""
+}