+++ docs/zh-cn/changelog.md
@@ -4,14 +4,14 @@
 
 > 发布时间：2021-08-08
 
-* 新增`tun`连接模式（alpha）
 * 增强Windows下的`connect`命令支持
+* 移除对本地`kubectl`客户端工具的依赖
+* 新增适用于Linux的`tun`连接模式（alpha）
 * 使用`provide`命令替代`run`命令
 * 新增`clean`命令，清理集群中残留的Shadow Pods
 * 支持`service.namespace.svc`结构的服务域名解析
 * 完善缺失`sshuttle`依赖等运行时错误的报错信息
-* dump2hosts支持完整服务域名
-* 移除对本地`kubectl`客户端工具的依赖
+* `connect`命令的`dump2hosts`参数支持完整服务域名
 
 ### 0.0.13-rc13